refactor(desktop): move chat domain modules to features/chat
turn, spawn-from-comment moved. all importers, vi.mock, and
dynamic imports updated including parallel-e2e test.
