+++ apps/desktop/src/features/chat/spawn-from-comment.ts
@@ -1,6 +1,6 @@
 import type { PrComment, PullRequestState } from '@kay-am/types';
-import type { AgentKind } from './features/session/agent-kind';
-import { AGENT_KIND_DEFAULTS } from './features/session/agent-kind';
+import type { AgentKind } from '../session/agent-kind';
+import { AGENT_KIND_DEFAULTS } from '../session/agent-kind';
 
 const TITLE_MAX = 60;
